Moved StreamMetadata to different package as suggested in review comment
