+++ pinot-core/src/main/java/com/linkedin/pinot/core/realtime/stream/StreamMessageDecoder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.linkedin.pinot.core.realtime.impl.kafka;
+package com.linkedin.pinot.core.realtime.stream;
 
 import java.util.Map;
